Move some types into their own files
