+++ FFXIVClientStructs/FFXIV/Common/Component/Excel/ExcelColumnType.cs
@@ -1,0 +1,25 @@
+﻿namespace FFXIVClientStructs.FFXIV.Common.Component.Excel;
+
+public enum ExcelColumnType : ushort {
+    String = 0x0,
+    Bool = 0x1,
+    Int8 = 0x2,
+    UInt8 = 0x3,
+    Int16 = 0x4,
+    UInt16 = 0x5,
+    Int32 = 0x6,
+    UInt32 = 0x7,
+
+    Float32 = 0x9,
+    Int64 = 0xA,
+    UInt64 = 0xB,
+
+    PackedBool0 = 0x19,
+    PackedBool1 = 0x1A,
+    PackedBool2 = 0x1B,
+    PackedBool3 = 0x1C,
+    PackedBool4 = 0x1D,
+    PackedBool5 = 0x1E,
+    PackedBool6 = 0x1F,
+    PackedBool7 = 0x20,
+}
