+++ FFXIVClientStructs/FFXIV/Common/Component/Excel/ExcelVariant.cs
@@ -1,0 +1,6 @@
+﻿namespace FFXIVClientStructs.FFXIV.Common.Component.Excel;
+
+public enum ExcelVariant : uint {
+    SingleRow,
+    MultiRow
+}
